poc/ckb_timer: add keyboard timer PoC
I have a Corsair K68 RGB keyboard here. I don't know exactly why, but
I'm typing on it now.

Using ckb-next, we can write commands to a pipe to change key colors.

First thing I'm going to use this for is to create a timer. It uses the
top row, the function keys, which all represent 300 seconds of time. 12
of them add up to 1 hour. Together with the print screen, scroll lock
and pause key, which are overflow keys for the next 3 hours in total, we
can visually represent the amount of time until a moment in the future.

Currently:
- the command pipe path is hardcoded
- amount of remaining seconds in total has to be provided as argument
- it only sets the colors once, it's not an actual timer yet...